fix: correcoes criticas de seguranca e bugs identificados na auditoria
- JwtUtils: int -> long em jwtRefreshExpirationMs (previne overflow futuro)
- AuthController: usar X-Real-IP em vez de X-Forwarded-For (previne IP spoofing no rate limiting)
- AuthContext: corrigir nome do evento ULTRA_BOOTING -> OMEGA_BOOTING (bug: indicador cold start nunca ativava)
- nginx.conf: restaurar redirecionamento HTTP->HTTPS, corrigir server_name para xavierbr-vps.tech, adicionar preload no HSTS
- docker-compose.prod.yml: passar APP_ROOT_PASSWORD e APP_ADMIN_PASSWORD para o backend
- deploy_omega.sh: abortar deploy se .env ausente ou variaveis criticas nao definidas
- application.properties: tornar CORS dinamico via variavel de ambiente
- .env.example: remover senha real de banco (substituida por placeholder)

+++ backend/src/main/java/com/portalcursos/ng02/security/JwtUtils.java
@@ -24,7 +24,7 @@ public class JwtUtils {
     private int jwtExpirationMs;
 
     @Value("${portalcursos.jwt.refresh-expiration}")
-    private int jwtRefreshExpirationMs;
+    private long jwtRefreshExpirationMs;
 
     public String generateTokenFromUsername(String username) {
         return Jwts.builder()
